[WIP] Proportional Font: Fix italic/bold widths
This fix issues rendering italic, bold, and italic+bord characters.
Both text, decoration, background and cursor position are fixed.
In order to compute the width of font variants, it's necessary
to add some fields in the `Font` struct -- one pointer to each of
the font metrics needed. For monospace fonts, these fields
are not used, and will only hold a `nil` pointer.