[WIP] Proportional Font: Always draw Single Rect
To avoid glitches when `IndentGuide` is set to false in configuration,
this commit force to `drawWithSingleRect` when updating window if a
proportional font is used.
This is a workaround to support deactivation of `IndentGuide` option.
It may be avoided be modifiying the function that calculate the sizes
of `rects` needing updating in the `Window.update()` method.