[WIP] Proportional Font: Completion windows
For the purpose of completion floating windows, every floating window
that has, in its config, `relative = "cursor"` or `relative = "editor"`
will be positioned relatively to the actual text on the buffer.
It should be a temporary solution until (I hope) neovim implement a way to
identify floating windows that are positioned near the cursor.
(See: https://github.com/neovim/neovim/issues/34595)
A consequence of this commit is that some floating window that must be
centered won't be centered anymore, since their X coordinate will
depend on the text that the buffer contained. But it may not be a real
issue, since the text itself, because of visual wrapping, doesn't occupy
the whole window.
(Note: as other commits, only proportional fonts are affected.)